Refactor: Centralize Logger and DatabaseManager, remove unused files
