--- public/data/rotation.config.json
@@ -1,75 +1,0 @@
-{
-  "schemaVersion": "2.0",
-  "cadenceDays": 7,
-  "rolloverUtc": "Monday 09:00",
-  "regions": {
-    "default": {
-      "cadenceDays": 7,
-      "rolloverUtc": "Monday 09:00"
-    },
-    "asia": {
-      "cadenceDays": 7,
-      "rolloverUtc": "Monday 02:00"
-    },
-    "europe": {
-      "cadenceDays": 7,
-      "rolloverUtc": "Monday 10:00"
-    }
-  },
-  "pools": {
-    "freeRotation": ["vanguard", "crusher", "phantom", "guardian"],
-    "starter": ["vanguard"],
-    "owned": [],
-    "eventOverrides": {
-      "2025-09-05": {
-        "add": ["tempest", "assassin"],
-        "remove": ["phantom"]
-      },
-      "2025-12-25": {
-        "add": ["crimson", "azure", "void"],
-        "remove": [],
-        "description": "Holiday Special Rotation"
-      }
-    }
-  },
-  "flags": {
-    "trainingAlwaysUnlocked": true,
-    "allowHotSwap": true,
-    "enableEventRotations": true,
-    "enableRegionalOverrides": true,
-    "debugMode": false
-  },
-  "analytics": {
-    "trackPickRate": true,
-    "trackWinRate": true,
-    "trackVariationUsage": true,
-    "trackRegionalDifferences": true
-  },
-  "entitlements": {
-    "devUnlocks": {
-      "enabled": true,
-      "characters": ["all"]
-    },
-    "qaFlags": {
-      "unlockAll": false,
-      "forceRotation": null
-    },
-    "platformSKUs": {
-      "seasonPass": {
-        "characters": ["dlc_pack_1", "dlc_pack_2"],
-        "enabled": true
-      },
-      "premiumEdition": {
-        "characters": ["all"],
-        "enabled": false
-      }
-    }
-  },
-  "ui": {
-    "showCountdown": true,
-    "showFeatured": true,
-    "featuredCharacters": ["vanguard"],
-    "lockIconStyle": "premium",
-    "rotationNotification": true
-  }
-}
