--- src/app/core/DatabaseManager.ts
@@ -1,26 +1,0 @@
-export class DatabaseManager {
-  private static instance: DatabaseManager;
-
-  private constructor() {
-    // private constructor for singleton
-  }
-
-  public static getInstance(): DatabaseManager {
-    if (!DatabaseManager.instance) {
-      DatabaseManager.instance = new DatabaseManager();
-    }
-    return DatabaseManager.instance;
-  }
-
-  public async initialize(_config: any): Promise<void> {
-    // Placeholder
-  }
-
-  public async checkHealth(): Promise<boolean> {
-    return true;
-  }
-
-  public async close(): Promise<void> {
-    // Placeholder
-  }
-}
