--- src/backend/identity/src/database/DatabaseManager.ts
@@ -1,32 +1,0 @@
-export class DatabaseManager {
-  private static instance: DatabaseManager;
-
-  private constructor() {
-    // private constructor for singleton
-  }
-
-  public static getInstance(): DatabaseManager {
-    if (!DatabaseManager.instance) {
-      DatabaseManager.instance = new DatabaseManager();
-    }
-    return DatabaseManager.instance;
-  }
-
-  public async initialize(config: any): Promise<void> {
-    // Placeholder
-  }
-
-  public async checkHealth(): Promise<boolean> {
-    return true;
-  }
-
-  public async close(): Promise<void> {
-    // Placeholder
-  }
-
-  async query(sql: string, params?: any[]): Promise<{ rows: any[] }> {
-    // Placeholder implementation - replace with actual database client
-    console.log('Database query:', sql, params);
-    return { rows: [] };
-  }
-}
